requirements.txt
anyascii==0.3.3
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.5
defusedxml==0.7.1
Django==6.0.3
django-filter==25.2
django-modelcluster==6.4.1
django-permissionedforms==0.1
django-stubs-ext==5.2.9
django-taggit==6.1.0
django-tasks==0.11.0
django-treebeard==4.8.0
djangorestframework==3.16.1
draftjs_exporter==5.2.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.11
laces==0.1.2
modelsearch==1.2
openpyxl==3.1.5
pillow==12.1.1
pillow_heif==1.3.0
requests==2.32.5
soupsieve==2.8.3
sqlparse==0.5.5
stripe==15.0.0
telepath==0.3.1
typing_extensions==4.15.0
urllib3==2.6.3
wagtail==7.3.1
Willow==1.12.0
